+++ CHANGELOG.md
@@ -1,9 +1,3 @@
-## [1.3.0]
-* Support flutter 2.2.0.
-
-## [1.2.2]
-* Checkbox supports tapping.
-
 ## [1.2.1]
 * Indented position not holding while editing.
