+++ lib/models/documents/document.dart
@@ -40,10 +40,6 @@ class Document {
 
   final Rules _rules = Rules.getInstance();
 
-  void setCustomRules(List<Rule> customRules) {
-    _rules.setCustomRules(customRules);
-  }
-
   final StreamController<Tuple3<Delta, Delta, ChangeSource>> _observer =
       StreamController.broadcast();
 
@@ -51,7 +47,7 @@ class Document {
 
   Stream<Tuple3<Delta, Delta, ChangeSource>> get changes => _observer.stream;
 
-  Delta insert(int index, Object? data, {int replaceLength = 0, bool autoAppendNewlineAfterImage = true}) {
+  Delta insert(int index, Object? data, {int replaceLength = 0}) {
     assert(index >= 0);
     assert(data is String || data is Embeddable);
     if (data is Embeddable) {
@@ -62,7 +58,7 @@ class Document {
 
     final delta = _rules.apply(RuleType.INSERT, this, index,
         data: data, len: replaceLength);
-    compose(delta, ChangeSource.LOCAL, autoAppendNewlineAfterImage: autoAppendNewlineAfterImage);
+    compose(delta, ChangeSource.LOCAL);
     return delta;
   }
 
@@ -75,7 +71,7 @@ class Document {
     return delta;
   }
 
-  Delta replace(int index, int len, Object? data, {bool autoAppendNewlineAfterImage = true}) {
+  Delta replace(int index, int len, Object? data) {
     assert(index >= 0);
     assert(data is String || data is Embeddable);
 
@@ -88,8 +84,7 @@ class Document {
     // We have to insert before applying delete rules
     // Otherwise delete would be operating on stale document snapshot.
     if (dataIsNotEmpty) {
-      delta = insert(index, data, replaceLength: len,
-          autoAppendNewlineAfterImage: autoAppendNewlineAfterImage);
+      delta = insert(index, data, replaceLength: len);
     }
 
     if (len > 0) {
@@ -129,13 +124,13 @@ class Document {
     return block.queryChild(res.offset, true);
   }
 
-  void compose(Delta delta, ChangeSource changeSource, {bool autoAppendNewlineAfterImage = true}) {
+  void compose(Delta delta, ChangeSource changeSource) {
     assert(!_observer.isClosed);
     delta.trim();
     assert(delta.isNotEmpty);
 
     var offset = 0;
-    delta = _transform(delta, autoAppendNewlineAfterImage: autoAppendNewlineAfterImage);
+    delta = _transform(delta);
     final originalDelta = toDelta();
     for (final op in delta.toList()) {
       final style =
@@ -179,28 +174,22 @@ class Document {
 
   bool get hasRedo => _history.hasRedo;
 
-  static Delta _transform(Delta delta, {bool autoAppendNewlineAfterImage = true}) {
+  static Delta _transform(Delta delta) {
     final res = Delta();
     final ops = delta.toList();
     for (var i = 0; i < ops.length; i++) {
       final op = ops[i];
       res.push(op);
-      if (autoAppendNewlineAfterImage) {
-        _autoAppendNewlineAfterImage(i, ops, op, res);
-      }
+      _handleImageInsert(i, ops, op, res);
     }
     return res;
   }
 
-  static void _autoAppendNewlineAfterImage(
+  static void _handleImageInsert(
       int i, List<Operation> ops, Operation op, Delta res) {
     final nextOpIsImage =
         i + 1 < ops.length && ops[i + 1].isInsert && ops[i + 1].data is! String;
-    if (nextOpIsImage &&
-        op.data is String &&
-        (op.data as String).isNotEmpty &&
-        !(op.data as String).endsWith('\n'))
-    {
+    if (nextOpIsImage && !(op.data as String).endsWith('\n')) {
       res.push(Operation.insert('\n'));
     }
     // Currently embed is equivalent to image and hence `is! String`
