+++ lib/models/rules/rule.dart
@@ -28,8 +28,6 @@ abstract class Rule {
 class Rules {
   Rules(this._rules);
 
-  List<Rule> _customRules = [];
-
   final List<Rule> _rules;
   static final Rules _instance = Rules([
     const FormatLinkAtCaretPositionRule(),
@@ -51,14 +49,10 @@ class Rules {
 
   static Rules getInstance() => _instance;
 
-  void setCustomRules(List<Rule> customRules) {
-    _customRules = customRules;
-  }
-
   Delta apply(RuleType ruleType, Document document, int index,
       {int? len, Object? data, Attribute? attribute}) {
     final delta = document.toDelta();
-    for (final rule in _customRules + _rules) {
+    for (final rule in _rules) {
       if (rule.type != ruleType) {
         continue;
       }
